requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
black==25.11.0
    # via -r requirements.in
build==1.3.0
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2025.11.12
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.4
    # via
    #   -r requirements.in
    #   requests
click==8.3.0
    # via
    #   -r requirements.in
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   -r requirements.in
    #   build
    #   click
    #   pytest
flake8==7.3.0
    # via -r requirements.in
idna==3.11
    # via
    #   -r requirements.in
    #   requests
iniconfig==2.3.0
    # via
    #   -r requirements.in
    #   pytest
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   pytest-html
markupsafe==3.0.3
    # via
    #   -r requirements.in
    #   jinja2
mccabe==0.7.0
    # via
    #   -r requirements.in
    #   flake8
mypy-extensions==1.1.0
    # via
    #   -r requirements.in
    #   black
packaging==25.0
    # via
    #   -r requirements.in
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via
    #   -r requirements.in
    #   black
pip-tools==7.5.2
    # via -r requirements.in
platformdirs==4.5.0
    # via
    #   -r requirements.in
    #   black
pluggy==1.6.0
    # via
    #   -r requirements.in
    #   pytest
pycodestyle==2.14.0
    # via
    #   -r requirements.in
    #   flake8
pyflakes==3.4.0
    # via
    #   -r requirements.in
    #   flake8
pygments==2.19.2
    # via
    #   -r requirements.in
    #   pytest
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
pytest==9.0.1
    # via
    #   -r requirements.in
    #   pytest-html
    #   pytest-metadata
pytest-html==4.1.1
    # via -r requirements.in
pytest-metadata==3.1.1
    # via
    #   -r requirements.in
    #   pytest-html
python-dotenv==1.2.1
    # via -r requirements.in
pytokens==0.3.0
    # via
    #   -r requirements.in
    #   black
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   requests
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
